third_party/WebKit/LayoutTests/http/tests/htmlimports/csp-import-block-but-unsafe-inline-and-eval-nested.html

<!DOCTYPE html>
<html>
<head>
<script>
</script>
<link id="target" rel="import" href="resources/child-having-unsafe-eval-policy.html">
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
shouldBeTrue("document.inlineScriptHasRun");
shouldBe("document.externalScriptHasRun", "undefined");
shouldBe("document.corsExternalScriptHasRun", "undefined");
shouldBeTrue("document.evalFromInlineHasRun");
shouldBe("document.evalFromExternalHasRun", "undefined");
shouldBe("document.evalFromCorsExternalHasRun", "undefined");
</script>
</body>
</html>
